backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
argon2-cffi==23.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Storage
boto3==1.34.0
django-storages==1.14.2

# API Documentation
drf-spectacular==0.26.5

# Payments
stripe==7.8.0

# Image Processing
Pillow==10.1.0

# Utilities
python-decouple==3.8
requests==2.31.0
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
